requirements.txt
streamlit
streamlit-webrtc
opencv-python
tensorflow
numpy
Pillow
